pyproject.toml
[tool.poetry]
name = "SubstituteBot"
version = "0.1.0"
description = """
This bot will allow you to create custom user groups (like \
whateveryoulike), assign users to them and use those groups in any \
onversation via inline mode. Those groups will be automatically expanded \
nto usernames of all their members. That way you can notify whoever you \
ike and they will surely hear you!\
"""
authors = ["Ilnur Garifullin <ilnur.email@gmail.com>", "Alexander Drovnyashin <s6040780@gmail.com>"]

[tool.poetry.dependencies]
python = "3.6.2"
python-telegram-bot = "^10.1"
python-decouple = "^3.1"
peewee = "^3.4"
transliterate = "^1.10"

[tool.poetry.dev-dependencies]
pytest = "^3.5"
